Add files via upload
This here, is the Calculator Application itself.
This is the most complex and practice project I ever planned. It was a simple start, with no big ideas, until I thought it could fit my GitHub, so I started overcomplicating things so much that, only with a lot of musical inspiration and ADHD moments, I could have put everything together. Many features may present unexpected behaviour, and I apologize, but some errors might include a bad use of the application (Not an insult, ok?).
So, the code itself have integrated CSS and JavaScript, which means, everything is kind of literally together. I was sure I was going to use In-Line coding at the moment I saw how to include a CSS file into the HTML main file. I never liked and still do not like any kind of Front-End coding, it is boring and uninteresting to me.
So JavaScript is also not my very interest, but I still can handle using it. It is actually pretty visible, take a look on how many lines of code I wrote in HTML, compared to the other two. My JavaScript knowledge is not as much as it seems, since most of the resources I used in the coding were learnt while I was coding, so I learnt while doing it.